merged-repos/auto-cropping/requirements.txt
altgraph==0.17.4
customtkinter
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
